pipeline/requirements.txt
httpx==0.28.1
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
schedule==1.2.2
